Crate matrix_sdk_crypto_ffi

Uniffi based bindings for the matrix-sdk-crypto crate.

This crate can be used to introduce E2EE support into an existing Matrix client or client library in any of the language targets Uniffi supports.

Structs

BackupKeys

Backup keys and information we load from the store.

BackupRecoveryKey

The private part of the backup key, the one used for recovery.

BootstrapCrossSigningResult

CancelInfo

Information on why a verification flow has been cancelled and by whom.

ConfirmVerificationResult

A result type for confirming verifications.

CrossSigningKeyExport

A struct containing private cross signing keys that can be backed up or uploaded to the secret store.

CrossSigningStatus

Struct representing the state of our private cross signing keys, it shows which private cross signing keys we have locally stored.

DecryptedEvent

An event that was successfully decrypted.

Device

An E2EE capable Matrix device.

DeviceLists

EncryptionSettings

Settings that should be used when a room key is shared.

KeyRequestPair

A pair of outgoing room key requests, both of those are sendToDevice requests.

KeysImportResult

MegolmV1BackupKey

The public part of the backup key.

MigrationData

Struct collecting data that is important to migrate to the rust-sdk

OlmMachine

A high level state machine that handles E2EE for Matrix.

PassphraseInfo

Struct containing info about the way the backup key got derived from a passphrase.

PickledAccount

A pickled version of an Account.

PickledInboundGroupSession

A pickled version of an InboundGroupSession.

PickledSession

A pickled version of a Session.

QrCode

The m.qr_code.scan.v1, m.qr_code.show.v1, and m.reciprocate.v1 verification flow.

RequestVerificationResult

A result type for requesting verifications.

RoomKeyCounts

Struct holding the number of room keys we have.

RoomSettings

Room encryption settings which are modified by state events or user options

Sas

The m.sas.v1 verification flow.

ScanResult

A result type for scanning QR codes.

SessionMigrationData

Struct collecting data that is important to migrate sessions to the rust-sdk

ShieldState

Take a look at [matrix_sdk_common::deserialized_responses::ShieldState] for more info.

SignatureUploadRequest

SignatureVerification

The result of a signature verification of a signed JSON object.

StartSasResult

A result type for starting SAS verifications.

UploadSigningKeysRequest

Verification

Enum representing the different verification flows we support.

VerificationRequest

The verificatoin request object which then can transition into some concrete verification method

VersionInfo

Build-time information about important crates that are used.

Enums

CryptoStoreError

DecodeError

Error type for the decoding and storing of the backup key.

DecryptionError

EventEncryptionAlgorithm

An encryption algorithm to be used to encrypt messages sent to a room.

HistoryVisibility

Who can see a room’s history.

KeyImportError

MigrationError

Error type for the migration process.

OutgoingVerificationRequest

PkDecryptionError

Error type for the decryption of backed up room keys.

QrCodeState

An Enum describing the state the QrCode verification is in.

Request

RequestType

SasState

An Enum describing the state the SAS verification is in.

SecretImportError

ShieldColor

Take a look at [matrix_sdk_common::deserialized_responses::ShieldState] for more info.

SignatureError

UserIdentity

Enum representing cross signing identities of our own user or some other user.

VerificationRequestState

An Enum describing the state the QrCode verification is in.

Traits

Logger

Trait that can be used to forward Rust logs over FFI to a language specific logger.

ProgressListener

Callback that will be passed over the FFI to report progress

QrCodeListener

Listener that will be passed over the FFI to report changes to a QrCode verification.

SasListener

Listener that will be passed over the FFI to report changes to a SAS verification.

VerificationRequestListener

Listener that will be passed over the FFI to report changes to a verification request.

Functions

migrate

Migrate a libolm based setup to a vodozemac based setup stored in a SQLite store.

migrate_room_settings

Migrate room settings, including room algorithm and whether to block untrusted devices from legacy store to Sqlite store.

migrate_sessions

Migrate sessions and group sessions of a libolm based setup to a vodozemac based setup stored in a SQLite store.

set_logger

Set the logger that should be used to forward Rust logs over FFI.