Struct matrix_sdk_crypto_ffi::QrCode

pub struct QrCode { /* private fields */ }

The m.qr_code.scan.v1, m.qr_code.show.v1, and m.reciprocate.v1 verification flow.

Implementations

impl QrCode

pub fn other_user_id(&self) -> String

Get the user id of the other side.

pub fn other_device_id(&self) -> String

Get the device ID of the other side.

pub fn flow_id(&self) -> String

Get the unique ID that identifies this QR code verification flow.

pub fn room_id(&self) -> Option<String>

Get the room id if the verification is happening inside a room.

pub fn is_done(&self) -> bool

Is the QR code verification done.

pub fn is_cancelled(&self) -> bool

Has the verification flow been cancelled.

pub fn we_started(&self) -> bool

Did we initiate the verification flow.

pub fn cancel_info(&self) -> Option<CancelInfo>

Get the CancelInfo of this QR code verification object.

Will be None if the flow has not been cancelled.

pub fn has_been_scanned(&self) -> bool

Has the QR verification been scanned by the other side.

When the verification object is in this state it’s required that the user confirms that the other side has scanned the QR code.

pub fn reciprocated(&self) -> bool

Have we successfully scanned the QR code and are able to send a reciprocation event.

pub fn cancel(&self, cancel_code: String) -> Option<OutgoingVerificationRequest>

Cancel the QR code verification using the given cancel code.

Arguments

cancel_code - The error code for why the verification was cancelled, manual cancellatio usually happens with m.user cancel code. The full list of cancel codes can be found in the spec

pub fn confirm(&self) -> Option<ConfirmVerificationResult>

Confirm a verification was successful.

This method should be called if we want to confirm that the other side has scanned our QR code.

pub fn generate_qr_code(&self) -> Option<String>

Generate data that should be encoded as a QR code.

This method should be called right before a QR code should be displayed, the returned data is base64 encoded (without padding) and needs to be decoded on the other side before it can be put through a QR code generator.

pub fn set_changes_listener(&self, listener: Box<dyn QrCodeListener>)

Set a listener for changes in the QrCode verification process.

The given callback will be called whenever the state changes.

pub fn state(&self) -> QrCodeState

Get the current state of the QrCode verification process.