Struct matrix_sdk_crypto_ffi::Sas

pub struct Sas { /* private fields */ }

The m.sas.v1 verification flow.

Implementations

impl Sas

pub fn other_user_id(&self) -> String

Get the user id of the other side.

pub fn other_device_id(&self) -> String

Get the device ID of the other side.

pub fn flow_id(&self) -> String

Get the unique ID that identifies this SAS verification flow.

pub fn room_id(&self) -> Option<String>

Get the room id if the verification is happening inside a room.

pub fn is_done(&self) -> bool

Is the SAS flow done.

pub fn we_started(&self) -> bool

Did we initiate the verification flow.

pub fn accept(&self) -> Option<OutgoingVerificationRequest>

Accept that we’re going forward with the short auth string verification.

pub fn confirm( &self ) -> Result<Option<ConfirmVerificationResult>, CryptoStoreError>

Confirm a verification was successful.

This method should be called if a short auth string should be confirmed as matching.

pub fn cancel(&self, cancel_code: String) -> Option<OutgoingVerificationRequest>

Cancel the SAS verification using the given cancel code.

Arguments

cancel_code - The error code for why the verification was cancelled, manual cancellatio usually happens with m.user cancel code. The full list of cancel codes can be found in the spec

pub fn get_emoji_indices(&self) -> Option<Vec<i32>>

Get a list of emoji indices of the emoji representation of the short auth string.

Note: A SAS verification needs to be started and in the presentable state for this to return the list of emoji indices, otherwise returns None.

pub fn get_decimals(&self) -> Option<Vec<i32>>

Get the decimal representation of the short auth string.

Note: A SAS verification needs to be started and in the presentable state for this to return the list of decimals, otherwise returns None.

pub fn set_changes_listener(&self, listener: Box<dyn SasListener>)

Set a listener for changes in the SAS verification process.

The given callback will be called whenever the state changes.

This method can be used to react to changes in the state of the verification process, or rather the method can be used to handle each step of the verification process.

This method will spawn a tokio task on the Rust side, once we reach the Done or Cancelled state, the task will stop listening for changes.

Flowchart

The flow of the verification process is pictured bellow. Please note that the process can be cancelled at each step of the process. Either side can cancel the process.

               ┌───────┐
               │Started│
               └───┬───┘
                   │
              ┌────⌄───┐
              │Accepted│
              └────┬───┘
                   │
           ┌───────⌄──────┐
           │Keys Exchanged│
           └───────┬──────┘
                   │
           ________⌄________
          ╱                 ╲       ┌─────────┐
         ╱   Does the short  ╲______│Cancelled│
         ╲ auth string match ╱ no   └─────────┘
          ╲_________________╱
                   │yes
                   │
              ┌────⌄────┐
              │Confirmed│
              └────┬────┘
                   │
               ┌───⌄───┐
               │  Done │
               └───────┘

pub fn state(&self) -> SasState

Get the current state of the SAS verification process.