Struct matrix_sdk_crypto_ffi::VerificationRequest

pub struct VerificationRequest { /* private fields */ }

The verificatoin request object which then can transition into some concrete verification method

Implementations

impl VerificationRequest

pub fn other_user_id(&self) -> String

The id of the other user that is participating in this verification request.

pub fn other_device_id(&self) -> Option<String>

The id of the other device that is participating in this verification.

pub fn flow_id(&self) -> String

Get the unique ID of this verification request

pub fn room_id(&self) -> Option<String>

Get the room id if the verification is happening inside a room.

pub fn is_done(&self) -> bool

Has the verification flow that was started with this request finished.

pub fn is_ready(&self) -> bool

Is the verification request ready to start a verification flow.

pub fn we_started(&self) -> bool

Did we initiate the verification request

pub fn is_passive(&self) -> bool

Has the verification request been answered by another device.

pub fn is_cancelled(&self) -> bool

Has the verification flow that been cancelled.

pub fn cancel_info(&self) -> Option<CancelInfo>

Get info about the cancellation if the verification request has been cancelled.

pub fn their_supported_methods(&self) -> Option<Vec<String>>

Get the supported verification methods of the other side.

Will be present only if the other side requested the verification or if we’re in the ready state.

pub fn our_supported_methods(&self) -> Option<Vec<String>>

Get our own supported verification methods that we advertised.

Will be present only we requested the verification or if we’re in the ready state.

pub fn accept( &self, methods: Vec<String> ) -> Option<OutgoingVerificationRequest>

Accept a verification requests that we share with the given user with the given flow id.

This will move the verification request into the ready state.

Arguments

user_id - The ID of the user for which we would like to accept the verification requests.

flow_id - The ID that uniquely identifies the verification flow.

methods - A list of verification methods that we want to advertise as supported.

pub fn cancel(&self) -> Option<OutgoingVerificationRequest>

Cancel a verification for the given user with the given flow id using the given cancel code.

pub fn start_sas_verification( &self ) -> Result<Option<StartSasResult>, CryptoStoreError>

Transition from a verification request into short auth string based verification.

Arguments

user_id - The ID of the user for which we would like to start the SAS verification.

flow_id - The ID of the verification request that initiated the verification flow.

pub fn start_qr_verification( &self ) -> Result<Option<Arc<QrCode>>, CryptoStoreError>

Transition from a verification request into QR code verification.

This method should be called when one wants to display a QR code so the other side can scan it and move the QR code verification forward.

Arguments

user_id - The ID of the user for which we would like to start the QR code verification.

flow_id - The ID of the verification request that initiated the verification flow.

pub fn scan_qr_code(&self, data: String) -> Option<ScanResult>

Pass data from a scanned QR code to an active verification request and transition into QR code verification.

This requires an active VerificationRequest to succeed, returns None if no VerificationRequest is found or if the QR code data is invalid.

Arguments

user_id - The ID of the user for which we would like to start the QR code verification.

flow_id - The ID of the verification request that initiated the verification flow.

data - The data that was extracted from the scanned QR code as an base64 encoded string, without padding.

pub fn set_changes_listener( &self, listener: Box<dyn VerificationRequestListener> )

Set a listener for changes in the verification request

The given callback will be called whenever the state changes.

pub fn state(&self) -> VerificationRequestState

Get the current state of the verification request.